[b00d4a9] VFS: remove inconsistency for Read/Write return type and parameter type, many added checks

+++ xbmc/cores/dvdplayer/DVDInputStreams/DVDInputStreamFile.cpp
@@ -118,7 +118,7 @@ int CDVDInputStreamFile::Read(BYTE* buf, int buf_size)
 {
   if(!m_pFile) return -1;
 
-  unsigned int ret = m_pFile->Read(buf, buf_size);
+  ssize_t ret = m_pFile->Read(buf, buf_size);
 
   /* we currently don't support non completing reads */
   if( ret <= 0 ) m_eof = true;
